Moved unit-tests into their own '/test/unit' directory
