+++ tests/unit/parserutils.test.ts
@@ -2,7 +2,7 @@
  * @file Jest tests for parserutils.ts
  */
 
-import * as ParserUtils from "../src/parserutils";
+import * as ParserUtils from "../../src/parserutils";
 
 test("isString", () => {
     expect(ParserUtils.isString(null)).toBe(false);
