+++ tests/unit/sequencer.test.ts
@@ -2,8 +2,8 @@
  * @file Jest tests for sequencer.ts
  */
 
-import * as Sequencer from "../src/sequencer";
-import { sleep } from "../src/utils";
+import * as Sequencer from "../../src/sequencer";
+import { sleep } from "../../src/utils";
 
 test("stopEndsTheUntilEndPromise", () => {
     const sequencer = Sequencer.createRunner();
