+++ tests/unit/tree.parser.test.ts
@@ -2,8 +2,8 @@
  * @file Jest tests for tree.parser.ts
  */
 
-import * as Tree from "../src/tree";
-import * as Tree̦Parser from "../src/tree.parser";
+import * as Tree from "../../src/tree";
+import * as Tree̦Parser from "../../src/tree.parser";
 
 test("fieldTypeIsParsedSuccessfully", () => {
     const json = `{
