+++ tests/unit/tree.path.test.ts
@@ -2,8 +2,8 @@
  * @file Jest tests for tree.path.ts
  */
 
-import * as Tree from "../src/tree";
-import * as TreePath from "../src/tree.path";
+import * as Tree from "../../src/tree";
+import * as TreePath from "../../src/tree.path";
 
 test("findPathToRoot", () => {
     let nodeToTest: Tree.INode | undefined;
