+++ tests/unit/treescheme.serializer.test.ts
@@ -2,9 +2,9 @@
  * @file Jest tests for treescheme.serializer.ts
  */
 
-import * as TreeSchemeParser from "../src/treescheme.parser";
-import * as TreeSchemeSerializer from "../src/treescheme.serializer";
-import * as Utils from "../src/utils";
+import * as TreeSchemeParser from "../../src/treescheme.parser";
+import * as TreeSchemeSerializer from "../../src/treescheme.serializer";
+import * as Utils from "../../src/utils";
 
 test("savedJsonIsIdenticalToReadJson", () => {
     const json = Utils.formatJson(`{
