+++ tests/unit/utils.test.ts
@@ -2,7 +2,7 @@
  * @file Jest tests for utils.ts
  */
 
-import * as Utils from "../src/utils";
+import * as Utils from "../../src/utils";
 
 test("withNewElement", () => {
     expect(Utils.withNewElement([1, 2, 3], 0, 1337)).toEqual([1337, 2, 3]);
